+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 - Fix continuous releases/assets delivery on tags builds #53
 - Attempt at making menubar hiding feature cross-platform #7 (comment)
 - Attempt at making "About Ferdi" consistent across platforms #47 (comment)
+- Attempt at fixing "Launch on start" feature on Linux #63
 - Add "Quick Switch" feature
+- Add "Scheduled Do-not-Disturb" feature
 
 # [5.3.3](https://github.com/meetfranz/franz/compare/v5.3.2...v5.3.3) (2019-09-18)
 - Merged Franz v5.3.3 into Ferdi